Renamed *.h to *.hpp
