+++ examples/infinite_horizon_lqr/main.cpp
@@ -1,8 +1,7 @@
-#include <robotics/robotics.h>
-
 #include <chrono>
 #include <iostream>
 #include <random>
+#include <robotics/robotics.hpp>
 
 int main()
 {
