+++ examples/infinite_horizon_lqr_plot/main.cpp
@@ -1,10 +1,10 @@
 #include <matplot/matplot.h>
-#include <robotics/robotics.h>
 
 #include <chrono>
 #include <iomanip>
 #include <iostream>
 #include <random>
+#include <robotics/robotics.hpp>
 
 int main()
 {
