--- include/robotics/classical-control/classical-control.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <robotics/classical-control/pid.h>
