+++ include/robotics/classical-control/classical-control.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <robotics/classical-control/pid.hpp>
