+++ include/robotics/common.hpp
