--- include/robotics/estimation/estimation.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <robotics/estimation/extended_kalman_filter.h>
