+++ include/robotics/estimation/estimation.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <robotics/estimation/extended_kalman_filter.hpp>
