--- include/robotics/linear-control/linear-control.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <robotics/linear-control/linear-quadratic-regulator.h>
