+++ include/robotics/linear-control/linear-control.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <robotics/linear-control/linear-quadratic-regulator.hpp>
