--- include/robotics/system/system.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include <robotics/system/linear-system.h>
-#include <robotics/system/nonlinear-system.h>
