+++ include/robotics/system/system.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include <robotics/system/linear-system.hpp>
+#include <robotics/system/nonlinear-system.hpp>
