+++ test/source/linear-control/linear-quadratic-regulator.cpp
@@ -1,7 +1,7 @@
 #include <doctest/doctest.h>
-#include <robotics/linear-control/linear-quadratic-regulator.h>
 
 #include <Eigen/Dense>
+#include <robotics/linear-control/linear-quadratic-regulator.hpp>
 #include <string>
 
 TEST_CASE("Robotics")
